#123 refactor: 스켈레톤 경로 변겅
