+++ src/components/@common/ProductCard.tsx
@@ -9,7 +9,7 @@ const ProductCard = ({ product }: { product: ProductTypes }) => {
     product;
   return (
     <Link href={`/${convertIdToCategory(categoryId)}/${id}`}>
-      <div className="min-w-30 flex flex-col items-center justify-center pt-[10px] lg:pb-[20px] md:pb-[15px] mobile:pb-[10px] px-[10px] mobile:w-full lg:max-w-[300px] mobile:gap-[10px] md:gap-[20px] gap-[25px] rounded-lg border-2 hover:border-main-blue border-gray-35 bg-gray-25">
+      <div className="flex flex-col items-center justify-start pt-[10px] lg:pb-[20px] md:pb-[15px] mobile:pb-[10px] px-[10px] h-full mobile:w-full lg:max-w-[300px] mobile:gap-[10px] md:gap-[20px] gap-[25px] rounded-lg border-2 hover:border-main-blue border-gray-35 bg-gray-25">
         <ImageComponent type="product" src={image} alt="이미지" />
 
         <div className="w-full flex flex-col gap-[10px] mobile:gap-[5px] lg:px-[20px] md:px-[16px] mobile:px-[10px]">
